Add method to handle default display name
Added test that checks the agency name is displayed when there is not a
display_name available. Added method to return the display name when
available; otherwise it returns the agency name. Updated the view to
call the agency display name method.

This is needed to ensure an agency name is being displayed on the main
page.