Display current agency in the title (#700)
Test added to check agency name appears in the title with the associated
step definition. Updated the page title to display the current agencies
name.

This is done to support multiple agencies.

The controller helper method current_agency cannot be called within the
context of a unit test. Stubbing the current_agency method during the
RSpec test to return an Agency.

Fixes AgileVentures/MetPlus_tracker#712